daemon/c_seccomp: Moved time related emulation code to seperate file
This moves the emulation code for the adjtimex(), clock_adjtime and
clock_settime() syscalls into the separate file 'c_seccomp/adjtimex.c'.
We did not change the emulation code itself. Just moved the corresponding
c_seccomp_emulate_adjtimex(), c_seccomp_emulate_adjtime(), and
c_seccomp_emulate_settime() helpers to the new file and generated necessary
interface code.
