+++ daemon/Makefile
@@ -149,7 +149,8 @@ endif
 SRC_CSECCOMP = \
 	c_seccomp/seccomp.c \
 	c_seccomp/mknod.c \
-	c_seccomp/init_module.c
+	c_seccomp/init_module.c \
+	c_seccomp/adjtimex.c
 
 SRC_CMODULES += \
 	c_net.c \
