+++ daemon/c_seccomp/seccomp.c
@@ -62,7 +62,6 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <sys/socket.h>
-#include <sys/timex.h>
 
 #include <linux/capability.h>
 #include <linux/filter.h>
@@ -309,161 +308,6 @@ c_seccomp_fetch_vm_new(c_seccomp_t *seccomp, int pid, void *rbuf, uint64_t size)
 	}
 
 	return lbuf;
-}
-
-static void
-c_seccomp_emulate_adjtime(c_seccomp_t *seccomp, struct seccomp_notif *req,
-			  struct seccomp_notif_resp *resp)
-{
-	int ret_adjtime = -1;
-	struct timex *timex = NULL;
-
-	if (!(COMPARTMENT_FLAG_SYSTEM_TIME & compartment_get_flags(seccomp->compartment))) {
-		DEBUG("Blocking call to SYS_clock_adjtime by PID %d", req->pid);
-		goto out;
-	}
-
-	DEBUG("Got clock_adjtime, clk_id: %lld, struct timex *: %p", req->data.args[0],
-	      (void *)req->data.args[1]);
-
-	// Check cap of target pid in its namespace
-	if (!c_seccomp_capable(req->pid, CAP_SYS_TIME)) {
-		ERROR("Missing CAP_SYS_TIME for process %d!", req->pid);
-		goto out;
-	}
-
-	if (CLOCK_REALTIME != req->data.args[0]) {
-		DEBUG("Attempt of container %s to execute clock_settime on clock %llx blocked",
-		      uuid_string(compartment_get_uuid(seccomp->compartment)), req->data.args[0]);
-		goto out;
-	}
-
-	if (!(timex = (struct timex *)c_seccomp_fetch_vm_new(
-		      seccomp, req->pid, (void *)req->data.args[1], sizeof(struct timex)))) {
-		ERROR_ERRNO("Failed to fetch struct timex");
-		goto out;
-	}
-
-	DEBUG("Executing clock_adjtime on behalf of container");
-	if (-1 == (ret_adjtime = clock_adjtime(CLOCK_REALTIME, timex))) {
-		audit_log_event(NULL, FSA, CMLD, CONTAINER_ISOLATION, "seccomp-emulation-failed",
-				compartment_get_name(seccomp->compartment), 2, "syscall",
-				SYS_clock_adjtime);
-		ERROR_ERRNO("Failed to execute clock_adjtime");
-		goto out;
-	}
-
-	DEBUG("clock_adjtime returned %d", ret_adjtime);
-
-	// prepare answer
-	resp->id = req->id;
-	resp->error = 0;
-	resp->val = ret_adjtime;
-
-out:
-	if (timex)
-		mem_free(timex);
-}
-
-static void
-c_seccomp_emulate_adjtimex(c_seccomp_t *seccomp, struct seccomp_notif *req,
-			   struct seccomp_notif_resp *resp)
-{
-	int ret_adjtimex = -1;
-	struct timex *timex = NULL;
-
-	if (!(COMPARTMENT_FLAG_SYSTEM_TIME & compartment_get_flags(seccomp->compartment))) {
-		DEBUG("Blocking call to SYS_adjtimex by PID %d", req->pid);
-		goto out;
-	}
-
-	DEBUG("Got adjtimex, struct timex *: %p", (void *)req->data.args[0]);
-
-	// Check cap of target pid in its namespace
-	if (!c_seccomp_capable(req->pid, CAP_SYS_TIME)) {
-		ERROR("Missing CAP_SYS_TIME for process %d!", req->pid);
-		goto out;
-	}
-
-	if (!(timex = (struct timex *)c_seccomp_fetch_vm_new(
-		      seccomp, req->pid, (void *)req->data.args[0], sizeof(struct timex)))) {
-		ERROR_ERRNO("Failed to fetch struct timex");
-		goto out;
-	}
-
-	DEBUG("Executing adjtimex on behalf of container");
-	if (-1 == (ret_adjtimex = adjtimex(timex))) {
-		audit_log_event(NULL, FSA, CMLD, CONTAINER_ISOLATION, "seccomp-emulation-failed",
-				compartment_get_name(seccomp->compartment), 2, "syscall",
-				SYS_adjtimex);
-		ERROR_ERRNO("Failed to execute adjtimex");
-		goto out;
-	}
-
-	DEBUG("adjtimex returned %d", ret_adjtimex);
-
-	// prepare answer
-	resp->id = req->id;
-	resp->error = 0;
-	resp->val = ret_adjtimex;
-
-out:
-	if (timex)
-		mem_free(timex);
-}
-
-static void
-c_seccomp_emulate_settime(c_seccomp_t *seccomp, struct seccomp_notif *req,
-			  struct seccomp_notif_resp *resp)
-{
-	int ret_settime = -1;
-	struct timespec *timespec = NULL;
-
-	if (!(COMPARTMENT_FLAG_SYSTEM_TIME & compartment_get_flags(seccomp->compartment))) {
-		DEBUG("Blocking call to SYS_clock_settime by PID %d", req->pid);
-		goto out;
-	}
-
-	DEBUG("Got clock_settime, clockid: %lld, struct timespec *: %p", req->data.args[0],
-	      (void *)req->data.args[1]);
-
-	// Check cap of target pid in its namespace
-	if (!c_seccomp_capable(req->pid, CAP_SYS_TIME)) {
-		ERROR("Missing CAP_SYS_TIME for process %d!", req->pid);
-		goto out;
-	}
-
-	if (CLOCK_REALTIME != req->data.args[0]) {
-		DEBUG("Attempt of container %s to execute clock_settime on clock %llx blocked",
-		      uuid_string(compartment_get_uuid(seccomp->compartment)), req->data.args[0]);
-		goto out;
-	}
-
-	if (!(timespec = (struct timespec *)c_seccomp_fetch_vm_new(
-		      seccomp, req->pid, (void *)req->data.args[1], sizeof(struct timespec)))) {
-		ERROR_ERRNO("Failed to fetch struct timespec");
-		goto out;
-	}
-
-	DEBUG("Executing clock_settime on behalf of container");
-	if (-1 == (ret_settime = clock_settime(CLOCK_REALTIME, timespec))) {
-		audit_log_event(NULL, FSA, CMLD, CONTAINER_ISOLATION, "seccomp-emulation-failed",
-				compartment_get_name(seccomp->compartment), 2, "syscall",
-				SYS_clock_settime);
-		ERROR_ERRNO("Failed to execute clock_settime");
-		goto out;
-	}
-
-	DEBUG("clock_settime returned %d", ret_settime);
-
-	// prepare answer
-	resp->id = req->id;
-	resp->error = 0;
-	resp->val = ret_settime;
-
-out:
-	if (timespec)
-		mem_free(timespec);
 }
 
 static void
